Update patch set 1
Patch Set 1:

Since the protos are specific to openOLT then I suggest to create a protos directory under the openolt adapter and add these protos files there.  You also want to have an openolt build_protos.sh script.   Protos that are used by all will be in higher level directories and those that are specific should be closer to the specialized directory.

Patch-set: 1