Update patch set 3
Patch Set 3:

> > Yes, but the point of this commit is to move that process out of
 > > voltha-go, and into voltha-protos.
 > >
 > > So that process no longer exists in voltha-go. It is built in
 > > voltha-protos, which is pulled by dep.
 > 
 > That's all fine but build artifacts shouldn't be stored in the
 > source repository but rather created at build time regardless of
 > where they reside. This approach is a recipe for dependency issues
 > and run-time failures that you'll never catch during the build when
 > the .pb file is out of date. An out-of-date .pb file bug may not
 > manifest itself for days/weeks/months in applications that use it
 > at run-time.

I understand, but everything I find about go, indicates that the actual built artifacts need to be made available to the time of being pulled. For example, dep ensure will not pull the proto stubs down unless they are already generated. Otherwise, the process fails at the time of dep ensure saying no go code was found in the import path. So I am forced to commit build artifacts like *.pb.go files and .pb. It also does not pull other files than the ones in the directories it imports, so all the scripts and proto files needed to build are not included.
We can choose not to commit them in the voltha-go repo. That just means anyone cloning will have to run dep ensure. 
If we include them, then there is some risk there if someone updates the protos without also rerunning dep ensure in the voltha-go and pushing that. So I would say we should not commit *.pb.go or *.pb in voltha-go, and require people to run dep ensure.

Patch-set: 3