Update patch set 1
Patch Set 1:

I think the issue with the non-stub files is dep was told not to prune any of them out because afrouter needed voltha.pb.   Im not sure if there is a way to selectively prune.   If not the need for voltha.pb, then just the stubs would be here.

Patch-set: 1