VOL-1489: Add uni port id metadata needed for olt flows
When the olt is asked to setup pon resources on the
onu/uni's behalf it needs to know what onu and uni port it
is dealing with for a particular flow.
Typically with flow decomposer the ports on the child
device (uni) are not visible on the parent. But in the case
of olt based flows (at least with BAL?) the parent device
needs to know the uni ports so it can create
gems, alloc id, and queues needed.

This patch adds new metadata containing the uni port
to the decomposed flows for use by the adapter.
This change also no longer removes METADATA passed
down from onos, keeping it intact.  It happens to contain
the uni port id in this particular flow's case.

This change also starts down a path of later changes
needed for tech profile given it too will be encoded
in metadata and that needs to make it into the adapter.
So this will likely be revisited.

-

Another change (a bug?) was that for downstream table based flows
(NNI to PON port) the flow from onos has the classifier match in port
as the NNI and the PON port as the outpu action.  But after decomposition
the classifer in port became the PON port and the action output the NNI,
effectively reversing the flow and making no sense.  This change
reverses the ingress/egress ports returned by the route graph.

Change-Id: I063e09cce75d168857f2aeac28b4c9f98dbbba43