VOL-1489: Add uni port id as tunnel id needed for olt flows
When the olt is asked to setup pon resources on the
onu/uni's behalf it needs to know what onu and uni port it
is dealing with for a particular flow.
Typically with flow decomposer the ports on the child
device (uni) are not visible on the parent. But in the case
of olt based flows (at least with BAL?) the parent device
needs to know the uni ports so it can create
gems, alloc id, and queues needed.

This patch adds new openflow tunnel id containing the uni port
to the decomposed flows for use by the adapter.

Change-Id: I0ea701b457ab5cb5877b953f58364d6a7806a58f