Update patch set 1
Patch Set 1: Code-Review-1

In 1.x, since the adapters were part of the Core then it was kind of ok to add a UNI port reference on the OLT device.  In 2.x, since the adapters are separate from the Core then having an adapter creating this reference in the Core model is not desirable as it increase the coupling between the two.  Besides, it will interfere with the flow decomposition logic as now it will have to differentiate between what an actual peer is and a "referenced" one.   Therefore, I do not support the peer reference logic.

Patch-set: 1
Reviewer: Gerrit User 1000056 <1000056@2a2bfe1b-c5c2-48ed-9ac8-16438ab24388>
Label: Code-Review=-1