[VOL-1586] Possible race condition during ONU discovery
Possible race condition in openolt python adapter during onu discovery.
We removed two go routines to wait until the AdoptDevice is finished before sending
response to ChildDeviceDetected to the openolt adapter. Also openolt is modified to yield
for it before sending ActivateONU to the olt_agent.
Tested a couple of times with BBSIM and there was no issue. Further advice needed if
there might be other side effects.

Change-Id: Ic01b114d717f5942d7140a60fcfa8f0bc2df8ce2