Update patch set 1
Patch Set 1:

> Recently about 3 out of 10 times i get a "no-route" and needed to
 > do the below to hack/force the update.  With this change
 > (especially given flowTableUpdated is gone) should hopefully remove
 > this need:
 > 
 > 
 > +++ b/rw_core/core/logical_device_agent.go
 > @@ -1009,6 +1009,9 @@ func (agent *LogicalDeviceAgent)
 > flowTableUpdated(args ...interface{}) interface
 > 
 > var groups *ofp.FlowGroups
 > lDevice, _ := agent.getLogicalDeviceWithoutLock()
 > +
 > +       agent.deviceGraph.ComputeRoutes(lDevice.Ports)
 > +
 > groups = lDevice.FlowGroups
 > 
 > 
 > Ill cherry pick and give it a try.

Yes, this should be gone now.   I tested it multiple times and things looks better.  However, since I tested it with ponsim and simulated olt these may not catch the condition you are seeing with openolt.    When I started seeing the no-route issue I was seeing it all the times.  The fix is in the device graph.

Patch-set: 1