Update patch set 1
Patch Set 1:

Im looking around here now a bit as well for the onu device adapter, testing the case where the pon fiber is unplugged.  enableAllPorts and also a disableAllPorts would be useful as the "onu down" indication is passed on (via interadapter message) to the onu so it can disable its uni ports.

I dont want to mess with this area if you guys are already here.  But a disableAllPorts() would very helpful and save on messaging too.

Patch-set: 1