Update patch set 1
Patch Set 1: Code-Review+2

(1 comment)

Can you please create a Jira for removing the default rules from the Core 2.0?  I am sure down the road a question would be asked about it and we will need the reason.  Having the reason documented in a Jira would be good to have.   Also, it will be good to specify what the same default rules behavior were working for Voltha 1.x but not in Voltha 2.x.

Thanks
Khen

Patch-set: 1