Update patch set 3
Patch Set 3:

> (1 comment)

It is sending alarms to kafka in this current state, actually this KafkaClient is passed as a parameter to the raise or clear alarm function of the concerned alarm type. It is used in line number 125 and 133.

Patch-set: 3
Reviewer: Gerrit User 1000564 <1000564@2a2bfe1b-c5c2-48ed-9ac8-16438ab24388>
Label: Code-Review=0