Update patch set 3
Patch Set 3:

(5 comments)

> Patch Set 3: Code-Review+1
> 
> > > (1 comment)
>  > >
>  > > Hi Matt,
>  > >
>  > > I see the content of the docker base being copied into several
>  > > dockerfiles.   Is there a reason?
>  > >
>  > > Thanks
>  > > Khen
>  > 
>  > See Zack's comment here.  https://gerrit.opencord.org/#/c/13889/
>  > It seems the automated build system needs either this sort of setup
>  > or base images having their own repo and versioning done.   I
>  > opt'ed for this in the short term.  Longer term voltha probably
>  > need a repo for a base golang and a base python image.
> 
> That's too bad.  I would have preferred the automated build system cater for this .

Using `base` as a tag to would have issues with trying to use any public build service (DockerHub, etc.).  Layering *is* possible, it just needs to be built with a unique name and version, but in my experience, sequencing of build order on public build services so the layers build in order has been problematic.

I made a few comments, and will need to fix one of the checker scripts - it's finding other Dockerfiles in `vendor/` that aren't related to our code.

Patch-set: 3