Update patch set 3
Patch Set 3: Code-Review+1

I gave this a +1 because the test/use case that was broke is lost and can't be reproduced. It is believed that the old hack is no longer needed, but it is not guaranteed. Reverting this will at least expose the flaw so that we can better track it if it does reappear.

Patch-set: 3
Reviewer: Gerrit User 1000003 <1000003@2a2bfe1b-c5c2-48ed-9ac8-16438ab24388>
Label: Code-Review=+1