Update patch set 6
Patch Set 6:

(2 comments)

> Patch Set 6:
> 
> kubectl has git commit plus "state" 
> 
> GitCommit:"3d12466c0221549ffd0651668e078b53482bcad2", GitTreeState:"dirty"
> 
> v
> 
> Commit:"3d12466c0221549ffd0651668e078b53482bcad2", GitTreeState:"clean"
> 
> to be honest i prefer this approach

As another point of reference - the label schema recommendation is to invalidate the `version`, but says nothing about `vcs-ref` when the tree is dirty: http://label-schema.org/rc1/

How the rest of CORD does it is to set `vcs-ref` to `unknown` if git tree is dirty.

Either David's approach (always have vcs-ref + clean/dirty flag) or the "rest of CORD" approach (vcs-ref changed `unknown` if dirty) is fine with me. 

In any case, the current patchset passes the semver check, but the `name` field needs to be manually filled in on the Dockerfiles (made notes of this on one of them)

Patch-set: 6
Label: Code-Review=0