Update patch set 7
Patch Set 7:

> > Can I please get an update on the comment I posted on May 10?
 > >
 > > Thanks
 > > Khen
 > Hi Khen, Dev has replied on May12, 12:12AM.


Thanks Girish.  My bad.  I was looking for comments on the file itself instead in the general section.

Once overall comment I have is that there is a distinction between an alarm and an event.  An event gets raised by the devices and can be classified as an alarm by a management system.   It would have been preferable to define this as the Event framework since this is about generating and publishing events.   

Thoughts?

Thanks
Khen

Patch-set: 7