Update patch set 7
Patch Set 7:

Is it ok if we put off the decision regarding DOCKER_LABEL_COMMIT_DATE and DOCKER_LABEL_VCS_REF or other alternatives and get this in?  This patch has already seen quite a bit of churn.

Also want to confirm again that folks are good with golang 1.12

Patch-set: 7
Label: Code-Review=0