Update patch set 1
Patch Set 1:

> The commit message mentions no need for entrypoint and cmd but I
 > see also entrypoint removed.  Is that intented?

Sorry, poorly worded. The intent is we don't need both and the other containers seem to use cmd so it can be easily overridden so we should only have cmd. This was noticed when i tried to use the container with the helm chart templates.

so we don't need "entrypoint and cmd" so instance just use "cmd"

Patch-set: 1