Update patch set 2
Patch Set 2:

Hello,

Can you please add more details to the Jira as it is unclear which issue is being addressed?   Looking at the code change it seems you are trying to prevent a parent device from disabling its child devices when the parent is disabled.  Is that correct?   Is this the expected behaviour?   The precondition for a device reboot is that the device should be disabled.  Is this the expected behaviour?  Maybe a discussion on the TST meeting would be useful.  I will send out an e-mail on this.

Khen

Patch-set: 2