Update patch set 2
Patch Set 2:

> > What happens if all containers never start? Is there an error
 > > returned to a GRPC call when all containers are not started that
 > > indicates the start state?
 > 
 > If all containers are not started, polling will simply continue
 > forever until they are.
 > 
 > > Would it be possible to add some test cases to the afrouter test
 > > case suite, specially for the scenario where a core crashes and
 > > rebalancing needs to occur?
 > 
 > In the next couple days, I plan to have k8s scheduling the pairs,
 > and afrouterd simply retrieving the existing pairing.  This will
 > remove the need for a complex repairing algorithm.

I think before refactoring the way the pairing work it is important to make sure that it works with the existing code base.  Hence, my comment about having  test cases for this use case.   This will identify what works as well as limitations.  That would be the baseline that the refactoring work can use.

Patch-set: 2
Label: Code-Review=0