Update patch set 2
Patch Set 2:

There are 2 variables that are gated.  One being the proxy attribute and the other is the one if for device ownership.  The proxy one might be overkill and could potentially be removed (will try).  Although, i don't see the risk.  It's a quick in/out operation for both SET and GET.  There are no other places that use that mutex and could cause a deadlock.  This part of the code has also been tested extensively.

Patch-set: 2
Reviewer: Gerrit User 1000119 <1000119@2a2bfe1b-c5c2-48ed-9ac8-16438ab24388>
Label: Code-Review=0