Update patch set 3
Patch Set 3:

> > > > Out of curiosity, is the afrouter in the non-ready state (k8s
 > > > ready
 > > > > probe) until all cores are started. It seems to me that it
 > > should
 > > > > be.
 > > >
 > > > Yes, that is one of the changes that this patch enables.
 > >
 > > What is the function/script that k8s needs to call as the ready
 > > probe?
 > 
 > Oh, I misunderstood your question.  The afrouterd doesn't start
 > until all the rw/ro cores are ready, but this isn't reported
 > anywhere (besides the logs).

well, it starts and as far as k8s is concerned it is "ready" which means the pod will have 2/2 containers ready and requests will flow to the pod. If we can someone have a probe that can be used to inform k8s that one or both of the containers is not ready then requests will not be forward to the pod and in the k8s management structures and operator can tell the pod isn't ready. 

A curl request into arouterd to determine if it is ready? is that possibld? a new JIRA?

Patch-set: 3
Reviewer: Gerrit User 1000003 <1000003@2a2bfe1b-c5c2-48ed-9ac8-16438ab24388>
Label: Code-Review=0