Update patch set 1
Patch Set 1:

There is a "UpdateMembership" API call by the api-server to the Cores (all of them) for each core to know to which core-pair it belongs.  The Core only uses it to setup a common topic for the pair.  Is this still the case with the changes your are making?

Patch-set: 1
Label: Verified=0