Update patch set 2
Patch Set 2:

Hi Mahir,

The reason "mergeDeviceInfoFromAdapter" is inside the adapter_request_handler.go is to limit the changes an Adapter can do to a device.   It is not meant to act as a gatekeeper for northbound request where a user can do pretty much what they want.   Moving it into the device_agent makes it a gatekeeper for northbound request as well.

Khen

Patch-set: 2