Update patch set 3
Patch Set 3:

> > Hi Khen,
 > > Then we have to find another way to get and update the device in
 > > same Lock block. Do you have  any  suggestion to do that?
 > 
 > All changes to a device is done with "Get" followed by an "Update"
 > within the same lock.   What do you mean by finding another way?
 > BTW, there is no northbound request to update an entire device.

"Get" is done in adapter_request_handler, in a lock, but "Update" is done in device_agent, in a separate Lock during DeviceUpdate. With this patch, I tried to fix this. But if the way I did it is not a preferred way(that is what I understand from your preview comment), then we should find a new way.

Patch-set: 3