Update patch set 5
Patch Set 5: Code-Review+1

> Hi Khen,
 > Changes were done. I tested with BBSIM with 300 ONUS. No deadlocks
 > are observed. What is the way you are testing? Is there another way
 > to test you suggest?

Hi Mahir,

We usually test in a 3-node cluster with 3 Core-pairs.   We create multiple OLTs and ONUs and enable them.   Given that you tested with 300 ONUs (I assume with 1 OLT and you enable all of them) and the nature of change then I am ok with it.   Are you using the python version of openolt or the Go version?

Khen

Patch-set: 5