Update patch set 15
Patch Set 15:

Could you run a dep ensure -update github.com/opencord/voltha-protos and push a new patchset?   There was an update to protos to have it match exactly voltha 1.7 and I think its good this patch set include it.

Im going through this now.  Im finding a few end-to-end issues, (BAL errors, onos apps not quite behaving right).  But honestly it will probably be easier to get this merged in and work through them once this and the openolt adapter are updated.  Ill write up what i find and we can discuss on the stabilization or regular voltha call.

Patch-set: 15
Label: Code-Review=0