Update patch set 5
Patch Set 5:

> (1 comment)

It depends on the context.  In this case, it would require another thread to listen on the channel & copy to the output stream.  Gothreads are lightweight, but in this case a mutex is much cheaper and has roughly the same throughput.

Patch-set: 5
Reviewer: Gerrit User 1000107 <1000107@2a2bfe1b-c5c2-48ed-9ac8-16438ab24388>
Label: Code-Review=0