Update patch set 2
Patch Set 2:

The best practice is to not merge a request until it has a +1 or greater from two individuals as well as the cherry-pick of the patchset is verified and ready to merge as well. if Khen can +1 based on the updated patchset that would be good for the two +1s. You also need to create a cherry pick to the voltha-2.1 branch and that needs to be verified by automated sanity tests.

Patch-set: 2