Update patch set 2
Patch Set 2:

This change seems like it's solving the the wrong problem.

It doesn't actually add a retry to the kv store connection.  Instead, it adds a delay that waits for the kv's own connection retry to succeed.

This adds a delay to the core startup (which waits for this to succeed before continuing).

The only "benefit" of this change is that the kv-store's probe status isn't set to "running" until it's connected.  However, it fails to update the probe status if connectivity is lost.

Proposal: add a kv-store state change callback in order to update the kv-store's probe value.  Don't delay starting of other cor components.

Also: Duplicate logic for RW Core.

Patch-set: 2
Label: Code-Review=0