Update patch set 21
Patch Set 21:

i pushed that image to dockerhub so k8s could find it and then started a pod with

kubectl run --image=dbainbriciena/test:foo test-app /app/test.py

this worked and i could exec into the pod and curl http://localhost:8911/

in short i am not clear why this would not work with the adapter.

that being said, i am not sure it matters, because you don't want to listen on localhost anyway. maybe default to "" and don't worry about localhost for now.

Patch-set: 21