[VOL-2174] Race condition when creating NNI port
This commit fixes the race condition where a the NNI logical port
is not created.  The root cause is that at the time an addport
request is received the parent id of the root device was not set
leadint the addport to believe that it does not exist even if it
does.

Change-Id: Ib9de9f4d3fd74fc142c35a75ba5f9a836985715b