Update patch set 7
Patch Set 7:

(4 comments)

The main question I have about this approach is there there is always a loop running as opposed to a loop that is only started when a failure is detected. What I was imagining, not that this is correct, is that the retry loop is only started after a failure is detected in some way. For example and wrt kafka, is a message send fails then kick off the retry loop.

Patch-set: 7
Label: Code-Review=0