Update patch set 2
Patch Set 2: Code-Review-1

(7 comments)

This is a good start. But I think the checks on the method calls need to check more than just no error, they need to verify the contents of the returned value v. expected. I commented this on some of the cases, but it needs to be done on all the cases. Additionally, negative tests need to be checked, e.g. delete something that doesn't exists or adding something that already exists.

Patch-set: 2
Label: Code-Review=-1