Update patch set 3
Patch Set 3:

> (1 comment)

It can be discussed. In my opinion it has to be skipped, otherwise gerrit stops. The developer could have modified the signature for any good reason and then merged the change. In any case it is tracked as "SKIP"ed and probably it is the test that has to be updated.

Patch-set: 3