Update patch set 2
Patch Set 2:

Nihil,

In general the changes you did are good.   You will need to manually rebase your changes.  Also, as you can see the unit tests for ro-core are failing.  Not sure if it's because of your update or a previous one.

Given the extent of ypur changes, can you run it in a clustered environment using kind-voltha?

Thanks
Khen

Patch-set: 2