Update patch set 4
Patch Set 4:

(2 comments)

This is much better, thank you. When you query back and get an agent, can we verify the values on the agent are what we expect (i.e. the ID is set correctly for example)? Are there other values on the returned value that we might verify?

Patch-set: 4