Update patch set 5
Patch Set 5: Code-Review-1

So, I'm opposed to this test overall.

There's no point testing the api, as it's entire purpose is to pass requests to some other component that actually does the work.  It's merely an interface that's implemented for gRPC.

I can probably be convinced otherwise, but at the moment, testing the api seems redundant.  Just test the various device managers that the api calls instead.

If you want to test the grpc_nbi_api_handler, test the KVtransaction and request ownership code that is unique to it, which is actually complicated enough to merit testing; and leave the rest alone.

Patch-set: 5
Label: Code-Review=-1