Update patch set 5
Patch Set 5:

(1 comment)

I do see value in the GetLogLevel/UpdateLogLevel tests, as there is logic in those methods in the handler (perhaps this illustrates the logic should have been moved elsewhere, but that's a slightly different point). Regarding the other functions, such as GetVoltha() I do see Kent's point, GetVoltha() is only a wrapper around a single function call. On the one hand, even a thing wrapper is code that ought to be tested, on the other hand it's so thin that it's hard to see a reasonable scenario where the test could fail. I think what to do in these thin wrappers is perhaps a discussion to have in channel.

Patch-set: 5
Label: Verified=0