Update patch set 8
Patch Set 8: Code-Review-1

I would argue that this test case is merely testing if a constant = itself.
  
Flags have a default value, but is testing that default value = the same default value really beneficial?

This also has the issue that it doesn't test new flags, or verify that the test is expanded to test new flags.  (adding a manual step to keep the test up to date, and not even verifying that that manual step is done.

I propose abandoning this test case, as it doesn't test anything of substance.

Patch-set: 8
Reviewer: Gerrit User 1000107 <1000107@2a2bfe1b-c5c2-48ed-9ac8-16438ab24388>
Label: Code-Review=-1