Update patch set 29
Patch Set 29: Code-Review-1

(25 comments)

Stopped adding comments as it was getting repetitive ;) Error checking needs to be more than log "error" and continue. If an error is returned processing need to stop and that error needs to be returned. If processing can continue it is not an error (it may be a warning). 

Ino most cases in this patchset it is an error and processing should stop and the error returned.

Patch-set: 29
Label: Code-Review=-1