Update patch set 2
Patch Set 2:

(1 comment)

Does the OLT Adapter (any OLT adapter implementation, not just OpenOLT)  pass down the flow "as-is" to the OLT device without decomposing it into an OLT and ONU flows?  In other words, no flows are required on the ONU, hence ONU adapter does not need to be in the picture?

Patch-set: 2
Label: Code-Review=0