Update patch set 14
Patch Set 14: Code-Review-1

(12 comments)

There is actually a significant amount of cleanup in the code required as part of the lint work. There are structures/interfaces/functions that are exposed outside of the package that should not be. An indicator of rework is that if you have to use `// nolint` changes are that rework is required.

Patch-set: 14
Label: Code-Review=-1