Update patch set 23
Patch Set 23:

> Why are GetEventFilter and ListEventFilters implemented only on the
 > ro_core?
 > 
 > Please duplicate the implementation to the rw_core (as the ro_core
 > is not in the current voltha stack, and will be removed.)

I was not aware of the fact. Also, I just kept the previous implementation with changes from alarm filter to event filter and planning to make modifications to it when event filtering is implemented. So if ro_core is going out then I can make the ListEventFilter and GetEventFilter as unimplemented in ro_core. Any thoughts?

Patch-set: 23
Label: Code-Review=0