Update patch set 2
Patch Set 2:

> > (2 comments)
 > >
 > > i think i understand a bit better. A walkthrough would be good
 > but
 > > i see no reason to hold this up.  Given the history and lack of
 > use
 > > of multicast a complete rewrite/removal of the processMulticastFlow
 > > function makes sense.
 > >
 > > I have some questions on the workflow, a sequence diagram may
 > help
 > > with each of the flows (group/bucket and trap) represented.
 > 
 > Hi Matt, I have added multicast setup diagram and some sequence
 > diagrams from Esin to Multicast implementation notes here
 > https://docs.google.com/document/d/1FNwkcDeVB3C2SpTJP_QrYTlC6sYFo_xwQnxASMk9x2U/edit#heading=h.ia3h0ht4tklh
 > . Please have look. Hope that helps.

It makes the OLT flows clearer.  It seems that indeed all the group/bucket work is done only on the OLT.  But the diagram does not show when or how the multicast ME are created on the ONU.  It seems that those will be setup before the group/bucket flows as part of the IGMP trap flow to the OLT.  But that would require IGMP trap flows to decompose or a new Interadapter message type to the onu.  I added a comment/question earlier in the doc on i suspect will be done, but need confirmation.

Patch-set: 2