Update patch set 2
Patch Set 2:

Just to clarify, there will actually be several "VoD" flows which will add vlans or pop them upstream/downstream. But the service definition in SADIS for VoD will include the "isIgmpRequired" field set to true. It is this field that will end up causing the OLT app to program the igmp flow in voltha, which will get decomposed and sent to both adaptors (as all traps are), and end up triggering ME creation.
 
And of course this has to happen before IGMP-join packets get trapped, and multicast flows/groups corresponding to the IGMP join are programmed

Patch-set: 2