WIP - Fix for jenkins NBI Failure
The root cause of the issue is a missing lock in the model.  Since
the change is quite involved this is a potential work around (it
may also be needed with the model change given that we are moving
away from core pairing and have a device being owned by one core at
all times.

Change-Id: Ic9e12d20c1ed41fed83acc8c29ae0150a014c406