WIP - Reworking interaction between the Core and Adapter
This commit is just to get feedback on Core/Adapter locking
mechanism rework.

There were two main issues raised regarding locks in the Core:

1) The Core locks a device agent until an Ack is received from
an Adapter.  In some cases this lock is held a long time
leading to timeouts on requests.

2) When the locks are held for a long time there can be multiple
requests for the same device waiting to acquire the lock.  Once
the lock is released any one of the request may end up acquiring
the lock where order of requests may not be respected.

This commit aim to address the above two issues.

Change-Id: I9b5d2063aaf91903a48dabdd109cc091ee72e795