[VOL-2454] Fix for device deletion in the rw_core
When a parent device is deleted, its child devices are deleted
as well.  The state machine in the Core ensures this process occurs.
However, there were "duplicate" codes during the parent device
deletion where it was triggering child devices in being deleted
again.  This is incorrect is being addressed by this change.

Change-Id: I8dbbe48e97f70638a0ea28444825452b366da137