[VOL-2442] Fix for Core panic
The logical device agent was receiving logical ports creation
while the logical device was not ready - it was waiting for
switch capability data from the OLT device.  This was causing
a panic.  The fix prevents logical port creation to be
trigerred when the logical device is not ready.  Once the
logical device is ready it will go over the ports data from the
OLT device and automatically create the logical ports.

Change-Id: Iad62302eda80fa158e59852810ad272a8aeedb7b