Update patch set 6
Patch Set 6:

> (1 comment)

I agree, but as part of this JIRA we only had to allow the core command line argument to take loglevel as string(and not int). 
Even with the existing implementation on specifying incorrect value, loglevel would be set to ErrorLevel. Can we create a new ticket for it?
As it would require changes in voltha-lib-go as well.

Patch-set: 6
Label: Code-Review=0