Update patch set 6
Patch Set 6:

I think it would make sense to have all the repeated code in profile.go and release.go in one place and whatever is truly different isolated.  Otherwise folks who are not interested or aware of profiling will change one file and not the other and cause confusion.

Patch-set: 6
Label: Code-Review=0