[VOL-2628] Fix issue with peer ports
There is a scenario where a request to add a PON peer port
is received before the OLT PON port has been created.  When this occurs
the relationship between the parent device and that child device is
lost.  Routes needed for flow decomposition cannot be created and
any request to retrieve child devices will fail.  This fix creates
a default PON port on the OLT when the peer request is received
and update the PON port with the latest info when the actual OLT
PON port creation is received.

Change-Id: I81807b3f4cbe7e9c1e3bbcb138fa8d4f20c6edeb