Update patch set 5
Patch Set 5: Code-Review+1

Khen, since this patch passes functional, hardware and also Matt's test it's not causing a regression, there might be things that coudl be optimised but I'd rather close this one which is long-lasting and open JIRA stories for optimisation. Also DT has a deadline to have this feature in and they will branch out if we do not merge it. I would advise putting this in today.

Patch-set: 5